backend/requirements.txt
# ============================================================================
# CORE FRAMEWORK
# ============================================================================
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
aiofiles>=23.2.1

# ============================================================================
# LANGCHAIN & LLM ECOSYSTEM
# ============================================================================
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
ollama>=0.3.0
groq>=0.11.0

# ============================================================================
# VECTOR DATABASE & EMBEDDINGS
# ============================================================================
chromadb>=0.5.0
sentence-transformers>=3.0.0
# Using stable torch compatible with Windows 
# NOTE: If you need a specific version (e.g. for CUDA), install this line manually first.
torch>=2.4.1 
torchvision>=0.19.1

# ============================================================================
# DOCUMENT PARSING & PROCESSING
# ============================================================================
pypdf>=4.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.23
markdown>=3.6

# ============================================================================
# OCR & IMAGE PROCESSING
# ============================================================================
pytesseract>=0.3.10
pillow>=11.0.0

# ============================================================================
# SEARCH & RANKING
# ============================================================================
rank-bm25>=0.2.2
scikit-learn>=1.5.0

# ============================================================================
# DATA PROCESSING
# ============================================================================
pandas>=2.2.0
numpy==1.26.4,<2.0.0 # Pinning numpy < 2.0.0

# ============================================================================
# DATABASE & ORM
# ============================================================================
sqlalchemy>=2.0.30
alembic>=1.13.0

# ============================================================================
# CONFIGURATION & VALIDATION
# ============================================================================
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1

# ============================================================================
# LOGGING & MONITORING
# ============================================================================
python-json-logger>=2.0.7

# ============================================================================
# UTILITIES
# ============================================================================
requests>=2.32.0
httpx>=0.27.0
tqdm>=4.66.0
python-slugify>=8.0.0

# ============================================================================
# TESTING
# ============================================================================
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0

# ============================================================================
# DEVELOPMENT TOOLS
# ============================================================================
black>=24.0.0
flake8>=7.1.1
isort>=5.13.2
mypy>=1.11.2
pylint>=3.3.1
pdf2image>=1.17.0

passlib[bcrypt]
PyJWT




